+++ aiohttp_sqlalchemy/web_handlers.py
@@ -6,6 +6,7 @@ from aiohttp.web import View
 from sqlalchemy import delete, select, update
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.sql import Delete, Select, Update
+from sqlalchemy_things.pagination import OffsetPaginator
 
 from aiohttp_sqlalchemy.constants import SA_DEFAULT_KEY
 from aiohttp_sqlalchemy.utils import get_session
@@ -35,6 +36,10 @@ class SAModelEditMixin(SAModelMixin):
 class SAModelViewMixin(SAModelMixin):
     def get_sa_select_stmt(self, model: Any = None) -> Select:
         return select(model or self.sa_model)
+
+
+class OffsetPagination(ahth.PaginationMixin):
+    paginator = OffsetPaginator()
 
 
 class PrimaryKeyMixin(ahth.PrimaryKeyMixin, SAModelMixin, metaclass=ABCMeta):
@@ -98,7 +103,6 @@ class ListEditMixin(ahth.ListMixin, SAModelEditMixin, metaclass=ABCMeta):
 
 class ListViewMixin(
     ahth.ListMixin,
-    ahth.PaginationMixin,
     SAModelViewMixin,
     metaclass=ABCMeta,
 ):